requirements.txt
# requirements_fixed.txt - UPDATED
# Core LangChain v1.x ecosystem (latest compatible)
langchain<0.3.0
langchain-core<0.3.0
langchain-community<0.3.0
# LLM Integrations (latest v1.x compatible)
langchain-groq
langchain-openai
openai>=1.0.0
groq>=0.14.0
# Other essential packages
streamlit>=1.52.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
pypdf>=4.0.0
python-docx>=1.1.0
python-dotenv>=1.0.0
tqdm>=4.66.0
numpy>=2.1.1
pyyaml>=6.0.0
extra-streamlit-components>=0.1.71
# OCR and Image Processing
pytesseract>=0.3.10
pypdfium2>=4.20.0
pdf2image>=1.16.0
Pillow>=10.0.0
# Fix for meta tensor error
# accelerate>=0.26.0  <-- Removed to fix meta tensor error
transformers>=4.36.0
